+++ NetIdMobileSdk/src/main/java/de/netid/mobile/sdk/api/NetIdErrorCode.kt
@@ -15,9 +15,7 @@
 package de.netid.mobile.sdk.api
 
 enum class NetIdErrorCode {
-    Timeout,
     NetworkError,
-    Uninitialized,
     JsonDeserializationError,
     InvalidDiscoveryDocument,
     Unknown,
@@ -34,5 +32,8 @@ enum class NetIdErrorCode {
     ClientError,
     StateMismatch,
     NoIdAppInstalled,
+    InvalidClient,
+    InvalidGrant,
+    UnsupportedGrantType,
     Other
 }