+++ NetIdMobileSdk/src/main/java/de/netid/mobile/sdk/model/PermissionReadResponse.kt
@@ -20,11 +20,14 @@ import kotlinx.serialization.encodeToString
 import kotlinx.serialization.json.Json
 
 @Serializable
-data class SubjectIdentifiers(
-    @SerialName("tpid")
-    val tpId: String?,
-    @SerialName("sync_id")
-    val syncId: String? = null
+data class PermissionReadResponse(
+    // Default to unknown if status received is not mapped in PermissionResponseFailureStatus
+    @SerialName("status_code")
+    val statusCode: PermissionResponseStatus = PermissionResponseStatus.UNKNOWN,
+    @SerialName("subject_identifiers")
+    val subjectIdentifiers: SubjectIdentifiers? =null,
+    @SerialName("netid_privacy_settings")
+    val netIdPrivacySettings: List<NetIdPrivacySetting>? = null
 ) {
     override fun toString(): String {
         val format = Json { encodeDefaults = false }