+++ NetIdMobileSdk/src/main/java/de/netid/mobile/sdk/model/PermissionUpdateResponse.kt
@@ -20,12 +20,12 @@ import kotlinx.serialization.encodeToString
 import kotlinx.serialization.json.Json
 
 @Serializable
-data class PermissionUpdateResponse(
+class PermissionUpdateResponse (
     @SerialName("subject_identifiers")
     val subjectIdentifiers: SubjectIdentifiers
-) {
-    override fun toString(): String {
+)  {
+   override fun toString(): String {
         val format = Json { encodeDefaults = false }
         return format.encodeToString(this)
-    }
+   }
 }