+++ NetIdMobileSdk/src/main/java/de/netid/mobile/sdk/ui/AuthorizationPermissionFragment.kt
@@ -104,7 +104,6 @@ class AuthorizationPermissionFragment(
                     openApp(app)
                 }
             } else {
-                authorizationIntent.setPackage("com.android.chrome")
                 resultLauncher.launch(authorizationIntent)
             }
         }